+++ CHANGELOG.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release] - 1971-11-03
 
 ### Added
-- 
+- Switch between table/data view button for ChartBlock
 
 ### Changed
-- 
+- Fix the margins of ChartBlock download buttons and the size of ChartBlock placeholder image.
+- Add ChartBlock to the simple article page
 
 ### Deprecated
 -