+++ CHANGELOG.md
@@ -9,10 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release] - 1971-11-03
 
 ### Added
-- 
+-
 
 ### Changed
-- 
+-
 
 ### Deprecated
 - 
@@ -26,10 +26,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - 
 
+
 ## [Upcoming release] - 1971-11-03
+
+### Added
+- Switch between table/data view button for ChartBlock
+- Source link for ChartBlock
+
+### Changed
+- Fix the margins of ChartBlock download buttons and the size of ChartBlock placeholder image.
+- Add ChartBlock to the simple article page
+
+### Deprecated
+- 
+
+### Removed
+- 
 
 ### Fixed
 - Fixed the placeholder in input field
+
+### Security
+- 
 
 
 ## [0.0.63] - 2022-03-09 (New Skip to main content component and bugfixes)