+++ CHANGELOG.md
@@ -30,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming release] - 1971-11-03
 
 ### Added
-- 
+- Switch between table/data view button for ChartBlock
+- Source link for ChartBlock
 
 ### Changed
 - Fix the margins of ChartBlock download buttons and the size of ChartBlock placeholder image.
@@ -38,8 +39,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change the params in searchNoResults to allow hiding of 'remove all filters' link
 - Updated styling on searchNoResults
 - Hide filters header if there are no results
-- Switch between table/data view button for ChartBlock
-- Source link for ChartBlock
 
 ### Deprecated
 -